move: validator는 service에서만 사용되므로 service 패키지 내부에 위치
