+++ src/main/java/com/peeerr/climbing/service/LikeService.java
@@ -8,7 +8,7 @@ import com.peeerr.climbing.exception.ErrorCode;
 import com.peeerr.climbing.repository.LikeRepository;
 import com.peeerr.climbing.repository.MemberRepository;
 import com.peeerr.climbing.repository.PostRepository;
-import com.peeerr.climbing.validator.LikeValidator;
+import com.peeerr.climbing.service.validator.LikeValidator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
